Short live QCoreApplication::instanceExists()
This is a thread-safe version of

    QCoreApplication::instance() != nullptr

for Qt 6.x, because QCoreApplication::self is not atomic and thus
reading it from outside the main thread could be a data race.

That's not to say it always is: if by construction the code can only run
in the main thread or while QCoreApplication definitely exists, that's
safe. Therefore, this commit focuses on places that are meant to be used
in multi-threaded environment (ruling out most of QtGui and QtWidgets)
or where the code was going to dereference the returned pointer anyway.

Change-Id: I6fc556c5fe5cbe0b5902fffdfb6b3bb345b0ee50
Reviewed-by: Marc Mutz <marc.mutz@qt.io>