QCoreApplication: avoid extra work in applicationFilePath() if we can
In the four OSes where qAppFileName() returns non-empty, the returned
path is always absolute and canonical, so there's no need for us to
further canonicalize it. So let's have applicationDirPath() have an
early out for it and mark it as Q_LIKELY().

In the other OSes, qAppFileName() is static inline in this very file, so
the compiler will probably see that it's always empty and ignore the
Q_LIKELY().

That leaves the atypical cases of /proc/self/exe being empty on Linux (I
don't know when this could happen, but it's not impossible) or non-
bundle executables on Apple OSes.

Change-Id: Ic47b28bda60085bcbcc8fffd28a0f3b864e07fd5
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>