rhi: vulkan: avoid allocations/deallocations in every beginPass() call
It is not uncommon to have more than 4 total clear values because of all
the color, depth, resolve and shading rate attachments.

Set the stack prealloc of the temporary QVarLengthArray to its maximum
possible size to avoid any heap (de)allocation in this hot path.

For 10 render passes per frame, it saves 600 allocs/deallocs per second.

Change-Id: If44fd075d1c1f9c07809ddc9922c690e9dc3ac3d
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>