QCborStreamReader: micro optimization: use constBegin() not constData()
Both QString and QByteArray's constData() have a check for null and
return the address of the corresponding _empty. constBegin() has no such
check and is therefore marginally faster because it could return null.

It will never in any of these conditions. Our internal buffer is never
empty at the places we're trying to get to its data; the params->string
and params->array have just been resized, so they are detached from null
even if newSize == 0. That leaves one case of a nullptr coming from the
user, in addData(), but QByteArray::append(nullptr, 0) works just fine
too.

Change-Id: I799e5c06638d393df793fffd502c25028c0f9c2e
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>