qEnvironmentVariableIntegerValue: remove arbitrary length limit
This commit removes all uses of getenv_s() (but not _wgetenv_s in
qEnvironmentVariable()), which is what was imposing the length
limit with qEnvironmentVariableIntValue(). We don't need to use
getenv_s() -- it's not more secure than getenv(). Microsoft could have
made them thread-safe, but the documentation for them[1] indicates they
are not.

[ChangeLog][QtCore][QtEnvironment] qEnvironmentVariableIntValue() is no
longer limited in the length of the number.

[1] https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/getenv-s-wgetenv-s

Change-Id: I95d97966c6da43b59a17fffdab2f904d78a62e9a
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>