Expose the Vulkan 1.4 core API
Non-trivial update this time. Recent vk.xml adds the option of listing
"vulkiansc" in the api attribute, which we did not handle before in
qvkgen for command elements at all, and was not treated as a
comma-separated list for feature elements. Fix these up.

Also add 1.4 in QVulkanDefaultInstance, so that 1.4 goes into
VkApplicationInfo.apiVersion, when available, incl. in Qt Quick.

[ChangeLog][QtGui] QVulkanFunctions and QVulkanDeviceFunctions are
updated to expose the Vulkan 1.4 API as long as Qt is built against
Vulkan 1.4 headers.

[ChangeLog][Third-Party Code] Updated vk.xml to 1.4.308.

Change-Id: I8bc4bb578575bcbb189286a26617cd0052241be5
Reviewed-by: Andy Nichols <andy.nichols@qt.io>