configure: Make it easier to translate arguments to CMake variables
To add a configure argument that is translated to a CMake variable, one
can now use in qt_cmdline.cmake files:
    qt_commandline_option(foo
        TYPE boolean
        CMAKE_VARIABLE QT_FOO
    )

It's not necessary anymore to add a translate_boolean_input call to
QtProcessConfigureArgs.cmake. This is especially useful if a non-qtbase
module wants to add such configure arguments.

This patch adds the CMAKE_VARIABLE argument for configure options of
TYPE boolean and string. An attempt to use CMAKE_VARIABLE with other
types will result in an error.

Task-number: QTBUG-123064
Change-Id: I4c341e57993b729d08327cea57cf7c8a05c96c0e
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>