QDirListing: add ExcludeOther
Which is the same as ExcludeSpecial (the latter is new in 6.9). This is
to match std::filesystem::is_other() which has a similar functionality.
This makes it easier to remember/understand both APIs.

Amends e583c3d5163a5512abac85e32359652e28a053f7.

[ChangeLog][QtCore][QDirListing] Add IteratorFlag::ExcludeOther
enumerator. IteratorFlag::ExcludeSpecial now becomes an, obsolete, alias
for ExcludeOther (the name was changed to "other" to match the wording
used in std::filesystem::is_other()).

Change-Id: I986a2d822615f69ccf5ec283756ae063f5e18101
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>