QSignalSpy: Drop the RegisterMethodArgumentMetaType metacall
There's nothing the meta-call could achieve: If the argument tpye was
complete when the meta-object was created, we would already have created
the metatype-interface for it, and would return a QMetaType constructed
from it in QMetaMethod::parameterMetatype. If the type wasn't complete,
parameterMetatype would have done the name to type lookup.

Consequently, if the lookup failed so far, calling
RegisterMethodArgumentMetaType would not help either, unless we would
have a dynamic meta-object with a custom implementation of
RegisterMethodArgumentMetaType. Such an implementation does not exist in
Qt.

This removes the last remaining user of RegisterMethodArgumentMetaType.

Change-Id: I9b9abbc3b4ba573b242027ecaa230ea27cb2ebf6
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>