QGIM: support gadgets and objects as rows and multi-role values
We can interpret the properties of a gadget or object in two ways:
either as different columns in a fixed-width table, or as different
role-values in a single item, matching the named properties against
the role names.

If used in a table structure, then the properties can only be
interpreted as different role values, as we already have a row type
with multiple columns.

In a list it can be either. To disambiguate in this case, check
whether there is a "display" property. If so, then we assume that
the gadget is supposed to be a multi-role item.

Change-Id: I99400304986f643506e590c5287f4d800654ec77
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>