QFactoryLoader: also move staticplugin instances to the main thread
We already move the instances coming from dynamic plugins, but for some
reason we don't for the ones coming from staticplugins. There's no hint
about it in the history; my guess is it was simply forgotten when the
staticplugins were added.

This necessitated extending the mutex lock over the staticplugin list
scan so we'd reach the QObject::moveToThread() call too.

Change-Id: Ibe0107a9daf46792b09efffd9a7d9c45214aa589
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>