CMake: Fix building tests with qt-cmake-standalone-test
We now require user projects to use find_package(Qt6FooPrivate) if
the Qt6::FooPrivate target is used. We don't require this for Qt's
autotests. But configuring such a test with qt-cmake-standalone-test
failed.

Fix this by setting QT_FIND_PRIVATE_MODULES to ON in the CMake project
template for standalone tests.

Change-Id: I778cbfc16638a94b5fe538c175394501df106b82
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>