Init QWidget::data to nullptr
It's unconditionally set to &d->data in d->init(), but at least
Coverity doesn't see it as that. Turn a potential uninit'ed value
access into a proper nullptr deref, in case Coverity has a point.

Not picking to older branches (at least not yet) in the unlikely case
that this unearths something.

Coverity-Id: 390797
Coverity-Id: 390790
Coverity-Id: 390785
Coverity-Id: 390783
Coverity-Id: 390775
Coverity-Id: 390774
Coverity-Id: 390771
Coverity-Id: 390765
Coverity-Id: 390762
Coverity-Id: 390751
Coverity-Id: 390737
Coverity-Id: 390734
Coverity-Id: 390727
Coverity-Id: 390724
Coverity-Id: 390707
Coverity-Id: 390706
Coverity-Id: 390700
Coverity-Id: 390693
Coverity-Id: 390692
Coverity-Id: 390690
Coverity-Id: 390684
Coverity-Id: 390680
Coverity-Id: 390677
Coverity-Id: 390675
Coverity-Id: 390669
Coverity-Id: 390662
Coverity-Id: 390660
Coverity-Id: 390659
Coverity-Id: 390657
Change-Id: Ieba218117d4bfdb4cdf5069211e9d12517b26d24
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>