QThread: disable pthread_clockjoin_np when building under TSAN
TSAN does not support intercepting pthread_clockjoin_np (only join,
tryjoin and timedjoin), and will claim that an application leaks threads.
Disable the clockjoin codepath when building under TSAN.

Unfortunately I can't seem to find a version check for the sanitizers to
apply here and know if and when TSAN will get this support.

Pick-to: 6.9
Change-Id: If12419e5d600e519bf4315d27f74d74a87dcf792
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>