QGIM: implement moveColumns
We can move columns when the row-type is a range (which guarantees
that all columns are of the same type), and if the columns are moved
within the same parent. Otherwise we would end up with different
branches of a tree having different column counts, which we don't allow.
Since we require trees to use statically sized row types, the former
implies the latter anyway.

Change-Id: Iaf9513d3642d6143cd4880937f2c022a0621c70d
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>