QGridLayout: make QHVContainer's preconditions machine-readable
Coverity complains that other() computes an address that's out of
m_data's bounds, and that's correct, if the argument isn't either
Qt::Horizontal or Qt::Vertical (they're flags, so Coverity assumes
that 0 and Vertical|Horizonal are valid arguments, too), but as the
doc block above the class suggests, this is not supported. So add
Q_ASSERTs to map() and mapOther(), and drop the noexcept of functions
that call these now.

Not picking to older branches in case Coverity is right and I am
wrong.

Amends 46a1cf915096ab056ad3bfd1d42fd504c04763cd.

Coverity-Id: 427374
Change-Id: I47a2821e99dd99666a170f2bb2859c1d6351b50e
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>